requirements.txt
weasyprint
jinja2
python-dotenv
google-genai
streamlit
PyMuPDF
python-docx
